Properly restore Core_unix dependency, needed for the manage dummy site
Core_unix seems to now be needed to access Caml_unix, which is used
for disabling terminal display when entering passwords (there is
a similar functionality available from Core_unix.Terminal_io but it is
essentially a matter of names).

We restore the dep by lack of an easier fix.